Move search to service
